refactor: 결과 판단 로직을 GameResult 정적 팩토리 메서드 내로 이동 및 test 생성
