+++ src/main/java/blackjack/domain/player/Player.java
@@ -19,7 +19,7 @@ public abstract class Player {
 
     public boolean isBust() {
         return cards.isBust();
-    };
+    }
 
     public int calculateScore() {
         return cards.calculateScore();
