+++ src/main/java/blackjack/domain/result/GameResult.java
@@ -3,6 +3,9 @@ package blackjack.domain.result;
 import blackjack.domain.player.Dealer;
 import blackjack.domain.player.Gamer;
 import blackjack.domain.player.Player;
+
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -14,8 +17,16 @@ public class GameResult {
         this.gameResult = gameResult;
     }
 
-    public static GameResult of(Map<Player, List<ResultType>> gameResult) {
-        return new GameResult(gameResult);
+    public static GameResult of(Dealer dealer, List<Gamer> gamers) {
+        Map<Player, List<ResultType>> result = new LinkedHashMap<>();
+        result.put(dealer, new ArrayList<>());
+        for (Gamer gamer : gamers) {
+            result.put(gamer, new ArrayList<>());
+            Map<Player, ResultType> resultPerGamer = ResultType.judgeGameResult(dealer, gamer);
+            result.get(gamer).add(resultPerGamer.get(gamer));
+            result.get(dealer).add(resultPerGamer.get(dealer));
+        }
+        return new GameResult(result);
     }
 
     public List<ResultType> findByPlayer(Player player) {
